+++ CHANGELOG.rst
@@ -3,8 +3,8 @@ Changelog
 
 `unreleased`_
 -------------
-* Added ``offline`` option to  ``make_azure_blueprint``. When set to ``True``,
-  automatic token refresh is enabled
+* Set ``auto_refresh_url`` automatically in ``make_azure_blueprint`` when the ``offline_access``
+  scope is included, thereby enabling automatic token refresh
 
 `7.0.1`_ (2024-01-05)
 ---------------------