+++ CHANGELOG.rst
@@ -3,6 +3,10 @@ Changelog
 
 `unreleased`_
 -------------
+nothing yet
+
+`7.1.0`_ (2024-03-05)
+---------------------
 * Set ``auto_refresh_url`` automatically in ``make_azure_blueprint`` when the ``offline_access``
   scope is included, thereby enabling automatic token refresh
 * Allow returning a custom response from a ``oauth_error`` signal handler.
@@ -409,7 +413,8 @@ Fixed
 .. _flit: https://flit.pypa.io/
 .. _PKCE: https://www.rfc-editor.org/rfc/rfc7636
 
-.. _unreleased: https://github.com/singingwolfboy/flask-dance/compare/v7.0.0...HEAD
+.. _unreleased: https://github.com/singingwolfboy/flask-dance/compare/v7.1.0...HEAD
+.. _7.0.1: https://github.com/singingwolfboy/flask-dance/compare/v7.0.1...v7.1.0
 .. _7.0.1: https://github.com/singingwolfboy/flask-dance/compare/v7.0.0...v7.0.1
 .. _7.0.0: https://github.com/singingwolfboy/flask-dance/compare/v6.2.0...v7.0.0
 .. _6.2.0: https://github.com/singingwolfboy/flask-dance/compare/v6.1.1...v6.2.0